(makefile-mode): Set local-write-file-hooks, not write-file-hooks.
Version 1.16.
(makefile-mode): Correct `make-variable-buffer-local' to
`make-local-variable'.  Set indent-tabs-mode to t.

(makefile-insert-macro-ref): Do not append a space
to the inserted macro reference.
(makefile-insert-macro-ref): Likewise if the reference is to
one of makefile-runtime-macros-list.
(makefile-format-macro-ref): If the name of the macro ref
starts with a '(' or '{' then the enclosing parentheses are not added.

(makefile-warn-suspicious-lines): New function.

Changed mode name to "Makefile" instead of "makefile".