(facemenu-active-faces): Replaces function `facemenu-discard-redundant-faces'.  This version, written by Simon Marshall, is faster and does not require optional argument for recursive re-entry.  New argument FRAME allows check to be done relative to face definitions in any frame.
(facemenu-unlisted-faces): Remove font-lock faces from
the default list.  The list of face names was out of sync; to
prevent this from happenning again I made font-lock.el, and other
packages that create "private" faces, put them on the list
themselves.  This should give them a better chance of being
updated when the packages are changed.