(dun-special-object): Floppy disk will melt in inventory or room, regardless of whether or not Stallman statuette is around.
(dun-examine):
You can examine objects in the jar without taking them out.

(dun-take):
You can take objects from the jar while you are on the bus.

(dun-dig):
Message from digging on the bus is the same as when you dig and don't
find anything.

(dun-climb):
No longer errors out of the game when argument to "climb" is invalid.

(dun-put):
You can now put things in the jar, even if you are on the bus.

(dun-special-move):
"In" or "Out" command tells you if you are already on or off the bus.

(dun-sauna-heat):
Changed "begin to sweat" to "are perspiring"
so that it makes sense whether you are heating up or cooling down.

(dun-help):
Changed author e-mail address, added web page.
Added hint for batch mode.

(*global*):
Fixed spelling of Presely in global object list.

(*global*):
Added coconuts, tank, and lake as objects that are recognized.

(*global*):
Added `slip' as another way of describing the paper,
and `chip' as another way of describing the CPU.

(*global*):
Upcase  abbreviations of directions in room descriptions.

(dun-login):
Fixed erroneous login message to better-describe ftp limitations.

(dun-rlogin):
Added error message if user tries to rlogin back to pokey.

(dun-load-d):
Fixed so that if restore file isn't found which in non-batch mode,
window will switch back to game.