Rafael Laboissiere <rafael@icp.inpg.fr> writes:
I was just annoyed with the fact that there is no way
to dynamically control the insertion accuracy.  I mean:
the variable `todo-insert-threshold' does the job, but
it is not very handy if one wants to mix the two
behaviors (bisection and "insert right here under the
cursor").

Therefore I did a quick hack in the function
`todo-insert-item'.  Now by giving a prefix argument to
the insert command (i.e. by typing "C-u i"), entries
are inserted exactly at the line where the cursor is.
It would be better to give the value of
`todo-insert-threshold' as a numeric argument of
`todo-insert-item' (like "M-8 i"), but it's too late
now for continuing to hack.