Known bug: Do not try to create a mode specific theme in latex-mode. It may fail. --> take this out, this was only in a beta version
  release date set to August 28.

I have moved "The Right Look: Colors and Fonts for Frames and Modes" to the Options chapter?
It's not so much something for Emacs veterans.

Also, the official URL for the aquamacs homepage is http://aquamacs.org -- NOT http://www.aquamacs.org. (See "Getting Help", and "Emacs: the Aquamacs distribution")

I added a few new shortcuts.

Those Aquamacs-specific "command keys" were wrong. Option is actually Meta, and only with the correct setting. It's pretty much never A-*.

"Aquamacs-specific variables that can be customized in the Options menu"
---> this sounds like there are entries in the Options menu for these variables. I reworded this as
"Aquamacs-specific variables that can be customized in the Aquamacs customization group"

Added some names to the "thank you" section.

Regarding TEXniscope: I have investigated this and found that using TEXniscope with Aquamacs is definitely not straight forward. I couldn't get the interaction to work with PDF, and the configuration needed some subtle changes, especially given our new name "Aquamacs Emacs" instead of "Emacs". If you'd like to recommend it in the manual, I think you'll need to at least create a Wiki pagedetailing the configuration (see my posting on the mailing list) and refer to the Wiki page from the manual. I have added a rudimentary entry - it'd be good if you would summarize it properly.