* vc-svn.el (vc-svn-parse-status): Recognize 'added, 'removed, and 'unregistered states.
* vc-hooks.el (vc-state): Added 'removed state to documentation.

* vc-hg (vc-hg-parse-status): Recognize 'added, 'removed,  and
'up-to-date state.  Deal with the possibility that C for clean
might change to "=" in the next point release.

* vc-git.el (vc-git-parse-status): Recognize 'removed.

* vc.el (header comment): Better description of new dir-state.
(vc-compatible-state): Defines which states are mutually
compatible; usable with 'edited as a test for which can be committed on.
(vc-dired-hook): Turn off undo in the VC-Dired buffer, a speed tweak.
(vc-default-dired-state-info): Display 'removed state.

* vc-bzr.el (vc-bzr-dir-state): Recognize 'added.