smart-remove-remaining-spaces: Fix an args-out-of-range error
When an entire buffer is marked and then deleted smart-delete-region
gets called with a 'from' argument of 1, with gets passed to as the
'pos' argument to smart-remove-remaining-spaces, which turns into a 0
for the second argument of buffer-substring-no-properties. Zero is
not a valid character position, which is what causes the error.

The error is prevented by guarding the second argument
buffer-substring-no-properties to be no less than point-min.

Signed-off-by: David Reitter <david.reitter@gmail.com>