Revise line wrapping menu.  Save line wrapping variables in custom-file
Line wrapping is now set in a submenu (in Options).  The settings of truncation, wrapping, word wrapping and line breaking (auto fill) apply
to the current buffer and, as a default, for future buffers.  There is also a setting to detect word wrap in text files, which
now, when this is off, use the global line wrapping setting.  When it is on, it detects the wrapping style of a buffer automatically,
but defaults to whatever is set with the customization variable `auto-word-wrap-default-function'.