subst-char-in-region signals consistent changes.
Currently, subst-char-in-region appears to signal inconsistent
ranges to before-change and after-change functions -- it signals
the maximal range of stop to before-change, but the actual range of
stop to after-change.

This fix results in a consistent behaviour, although it is currently
overly conservative. It signals the maximal range both before and
after.