Remove some undo-boundary insertion logic.
This change removes logc that inserts an undo boundary automatically if
a change has occurred in another buffer. It is not clear what the
motivation for that logic was. It require specifically disabling during
a function call to undo_outer_limit_function (which modifies the
minibuffer). Uses of post-command-hook or after-change-functions which
alter other buffers result in significant changes to the undo-behaviour
in the current-buffer, which is counter-intuitive, and is hard to
disable from lisp since last_undo_buffer is not visible form lisp.