Undo delayed setting of representedFilename
Back out change d9d383, which delayed setting of the filename
to help address Bug#18757 (or similar problems.)

Fixes a bug where second frames showing a different file did
not get a file icon, as sendEvent was not (no longer) called
for each frame right after the filename was set in the
global/static variable introduced to pass the filename.

No flicker observed.