Attempt at finishing script to add gnutls libraries to bundle.
1. Updated the build-homebrew-libraries.sh script
2. Removed references to it from Makefile.in
3. Updated the minimum Mac OS X version to 10.11.
4. Added some notes to the changelog (and please review for appropriateness)

I think it should work reasonably well for release builds. It requires
an update to homebrew to fix a homebrew bug that was getting in the
way, but that fix is in the distributed version now.

Some other notes:

1. You can hook in the build-homebrew-libraries.sh script anywhere you
want at the end of the build process. The usage is

   sh aquamacs/build/build-homebrew-libraries.sh <path to application bundle> [<min-version>]

   From the top-level aquamacs directory used for running
   build-aquamacs, the path would normally be
   nextstep/Aquamacs.app. The minimum version defaults to 10.11.

2. Anyone building Aquamacs on their own should probably have gnutls
available now, whether through homebrew or not. They don't need to run
the script, since it will just work with the local version and they
presumably won't be distributing it.