Change build process for better compatibility on El Capitan.
- Updated changelog recent changes.
- Set build-script to stop on error and always save output to log.
- Add compiler/linker options to fail on incompatible OS versions.
- Configure now checks more carefully for compatible library functions.
- Set exceptions on compatibility version check for mkostemp.

When we turned on strict checking for compatibility with old
libraries, clang flags mkostemp() as incompatible as of
10.11. However, it's implemented in gnulib, which is the one we
end up using anyway. So the calls to it are marked with a pragma that
overrides the command-line flag to throw an error.

In nsterm.m, a similar trick is used to allow compiling code that
checks for a method that is undefined in 10.11. It is still safe
because of the runtime check.