Remove the bundled alternative python-mode
Over time, there has been more than one python-mode for
Emacs. In the early days of Aquamacs, there was an independent one
that was much superior to the one bundled with Emacs. Since then,
main Emacs has included a newer one.

It's easy to install the alternative one using the MELPA package
python-mode, so Aquamacs is no longer bundling or recommending a
different python mode.